fix: add refresh to pairing modal QR code (#954)
